\name{gitter-package}
\alias{Rgitter}
%- \alias{gitter}
\docType{package}
\title{
gitter
}
\description{
gitter is a image analysis package for R. Its primary goal is to allow users to robustly grid and quantify images of pinned colonies. It includes several image pre-processing techniques, such as autorotation of plates, noise removal, contrast adjustment and image resizing.
}
\details{
\tabular{ll}{
Package: \tab gitter\cr
Type: \tab Package\cr
Version: \tab 1.0 beta\cr
Date: \tab 2013-05-16\cr
License: \tab LGPL\cr
}

}
\author{
Omar Wagih <omar.wagih@utoronto.ca>
Leopold Parts <leopold.parts@gmail.com>
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
