% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Main.R
\name{gitter.demo}
\alias{gitter.demo}
\title{Run a demo of gitter}
\usage{
gitter.demo(eg = 1)
}
\arguments{
\item{eg}{Type of demo. 1 for a single image demo, 1 for a single image demo using a reference image. Default is 1.}
}
\description{
This function will run a demo of gitter.
}
\examples{
# gitter.demo()
}

