% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.r
\name{gitter.read}
\alias{gitter.read}
\title{Read in a data file as a \code{\link{gitter}} data object.}
\usage{
gitter.read(path)
}
\arguments{
\item{path}{Path to the data file generated by \code{\link{gitter}}.}
}
\value{
\code{\link{gitter}} \code{\link{data.frame}} object for use with plot, summary and warning functions.
}
\description{
This function will take a path to a data file generated by \code{\link{gitter}} and read it into a \code{\link{gitter}} object for use with plot, summary and warning functions.
}
\examples{
# Get dat file path
f = system.file("extdata", "sample.jpg.dat", package="gitter")
# Read in path as a gitter data object
g = gitter.read(f)
}
\keyword{dat}
\keyword{file}
\keyword{gitter}
\keyword{read}

