\name{gitter-package}
\alias{Rgitter}
%- \alias{gitter}
\docType{package}
\title{
gitter
}
\description{
gitter is a image analysis package for R. Its primary goal is to allow users to robustly and quickly grid and quantify biological images of pinned colonies. It includes several image pre-processing techniques, such as autorotation of plates, noise removal, contrast adjustment and image resizing. 
}
\details{
\tabular{ll}{
Package: \tab gitter\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-05-16\cr
License: \tab LGPL\cr
}

}
\author{
Omar Wagih \email{omar.wagih@utoronto.ca}, 
Leopold Parts \email{leopold.parts@gmail.com}
}
%\references{

%}
\keyword{ package }
%\seealso{
%}
%\examples{
%}
